Split workspace model parity gates
